chore(xen-api): modularize
