+++ packages/xen-api/src/_debug.js
@@ -1,0 +1,3 @@
+import debug from 'debug'
+
+export default debug('xen-api')
