+++ packages/xen-api/src/_isGetAllRecordsMethod.js
@@ -1,0 +1,3 @@
+const SUFFIX = '.get_all_records'
+
+export default method => method.endsWith(SUFFIX)
