+++ packages/xen-api/src/_isOpaqueRef.js
@@ -1,0 +1,3 @@
+const PREFIX = 'OpaqueRef:'
+
+export default value => typeof value === 'string' && value.startsWith(PREFIX)
