+++ packages/xen-api/src/_makeCallSetting.js
@@ -1,0 +1,8 @@
+export default (setting, defaultValue) =>
+  setting === undefined
+    ? () => defaultValue
+    : typeof setting === 'function'
+    ? setting
+    : typeof setting === 'object'
+    ? method => setting[method] ?? setting['*'] ?? defaultValue
+    : () => setting
