+++ packages/xen-api/src/transports/_UnsupportedTransport.js
@@ -1,0 +1,3 @@
+import makeError from 'make-error'
+
+export default makeError('UnsupportedTransport')
