--- packages/xen-api/src/transports/_utils.js
@@ -1,3 +1,0 @@
-import makeError from 'make-error'
-
-export const UnsupportedTransport = makeError('UnsupportedTransport')
