+++ packages/xen-api/src/transports/auto.js
@@ -1,7 +1,7 @@
 import jsonRpc from './json-rpc'
+import UnsupportedTransport from './_UnsupportedTransport'
 import xmlRpc from './xml-rpc'
 import xmlRpcJson from './xml-rpc-json'
-import { UnsupportedTransport } from './_utils'
 
 const factories = [jsonRpc, xmlRpcJson, xmlRpc]
 const { length } = factories
