+++ packages/xen-api/src/transports/xml-rpc-json.js
@@ -1,7 +1,8 @@
 import { createClient, createSecureClient } from 'xmlrpc'
 import { promisify } from 'promise-toolbox'
 
-import { UnsupportedTransport } from './_utils'
+import prepareXmlRpcParams from './_prepareXmlRpcParams'
+import UnsupportedTransport from './_UnsupportedTransport'
 
 const logError = error => {
   if (error.res) {
@@ -71,10 +72,7 @@ const parseResult = result => {
   throw new UnsupportedTransport()
 }
 
-export default ({
-  allowUnauthorized,
-  url: { hostname, path, port, protocol },
-}) => {
+export default ({ allowUnauthorized, url: { hostname, port, protocol } }) => {
   const client = (protocol === 'https:' ? createSecureClient : createClient)({
     host: hostname,
     path: '/json',
@@ -83,5 +81,6 @@ export default ({
   })
   const call = promisify(client.methodCall, client)
 
-  return (method, args) => call(method, args).then(parseResult, logError)
+  return (method, args) =>
+    call(method, prepareXmlRpcParams(args)).then(parseResult, logError)
 }
